Refactoring: move some functions in crop_mesh
Move some implementation functions from crop_mesh_2d.py to
merge_mesh.py, for usage in other nodes.
