+++ nodes/CAD/crop_mesh_2d.py
@@ -12,70 +12,14 @@ from mathutils import Vector
 
 from sverchok.node_tree import SverchCustomTreeNode, throttled
 from sverchok.data_structure import updateNode, no_space
-from sverchok.utils.geom_2d.merge_mesh import crop_mesh, crop_edges
-from sverchok.utils.geom_2d.lin_alg import is_ccw_polygon
+from sverchok.utils.geom_2d.merge_mesh import crop_mesh, crop_edges, crop_mesh_delaunay
+from sverchok.utils.decorators import deprecated
 
-try:
-    from mathutils.geometry import delaunay_2d_cdt as bl_crop_mesh
-except ImportError:
-    bl_crop_mesh = None
-
-
+@deprecated("Use sverchok.utils.geom_2d.merge_mesh.crop_mesh_delaunay")
 def get_bl_crop_mesh_faces(verts, faces, verts_crop, faces_crop, mode, epsilon):
-    """
-    Crop given faces by another set of faces by Blender internal function
-    :param verts: list of Sv vertices
-    :param faces: list of faces wich should be cropped
-    :param verts_crop: list of Sv vertices
-    :param faces_crop: list of faces by which mesh should be cropped
-    :param mode: 'inner' or 'outer'
-    :param epsilon: float, for nearness tests
-    :return: list of Sv vertices, faces, indexes of old faces
-    """
-    faces = [f if is_ccw_polygon([verts[i] for i in f], accuracy=epsilon) else f[::-1] for f in faces]
-    faces_crop = [f if is_ccw_polygon([verts_crop[i] for i in f], accuracy=epsilon) else f[::-1] for f in faces_crop]
-    merged_verts, merged_faces, faces_indexes, faces_crop_indexes = join_meshes(verts, faces, verts_crop, faces_crop)
-    merged_verts = [Vector(co[:2]) for co in merged_verts]
-    verts_new, _, faces_new, _, _, face_indexes = bl_crop_mesh(merged_verts, [], merged_faces, 3, epsilon)
-    if mode == 'inner':
-        faces_out = []
-        faces_index_out = []
-        for f, fi in zip(faces_new, face_indexes):
-            if not fi:
-                # it means new faces was generated
-                continue
-            in_1 = False
-            in_2 = False
-            for i in fi:
-                if i in faces_indexes:
-                    in_1 = True
-                else:
-                    in_2 = True
-                if in_1 and in_2:
-                    faces_out.append(f)
-                    faces_index_out.append(min(fi))
-                    break
-        verts_out, faces_out = del_loose(verts_new, faces_out)
-        return [v.to_3d()[:] for v in verts_out], faces_out, faces_index_out
-    else:
-        faces_out = []
-        faces_index_out = []
-        for f, fi in zip(faces_new, face_indexes):
-            if not fi:
-                # it means new faces was generated
-                continue
-            in_2 = False
-            for i in fi:
-                if i in faces_crop_indexes:
-                    in_2 = True
-                    break
-            if not in_2:
-                faces_out.append(f)
-                faces_index_out.append(min(fi))
-        verts_out, faces_out = del_loose(verts_new, faces_out)
-        return [v.to_3d()[:] for v in verts_out], faces_out, faces_index_out
+    return crop_mesh_delaunay(verts, faces, verts_crop, faces_crop, mode, epsilon)
 
-
+@deprecated("Does not work properly")
 def get_bl_crop_mesh_edges(verts, edges, verts_crop, faces_crop, mode, epsilon):
     # well, it does not work, delaunay function does not give enough information
     # about location mesh relatively to each other. Can be implemented with help of some another algorithm
@@ -83,26 +27,6 @@ def get_bl_crop_mesh_edges(verts, edges, verts_crop, faces_crop, mode, epsilon):
     m_verts = [Vector(co[:2]) for co in m_verts]
     verts_new, edges_new, faces_new, _, _, _ = bl_crop_mesh(m_verts, m_edges, m_faces, 2, epsilon)
     return [v.to_3d() for v in verts_new], edges_new
-
-
-def join_meshes(verts1, faces1, verts2, faces2):
-    faces_out = faces1 + [[i + len(verts1) for i in f] for f in faces2]
-    faces1_indexes = {i for i in range(len(faces1))}
-    faces2_indexes = {i + len(faces1) for i in range(len(faces2))}
-    return verts1 + verts2, faces_out, faces1_indexes, faces2_indexes
-
-
-def join_meshes2(verts1, edges1, verts2, faces2):
-    faces_out = [[i + len(verts1) for i in f] for f in faces2]
-    return verts1 + verts2, edges1, faces_out
-
-
-def del_loose(verts, poly_edge):
-    indx = set(chain.from_iterable(poly_edge))
-    verts_out = [v for i, v in enumerate(verts) if i in indx]
-    v_index = dict([(j, i) for i, j in enumerate(sorted(indx))])
-    poly_edge_out = [list(map(lambda n: v_index[n], p)) for p in poly_edge]
-    return verts_out, poly_edge_out
 
 
 class SvCropMesh2D(bpy.types.Node, SverchCustomTreeNode):
@@ -173,7 +97,7 @@ class SvCropMesh2D(bpy.types.Node, SverchCustomTreeNode):
     def process(self):
         if not all([sock.is_linked for sock in self.inputs]):
             return
-        if self.alg_mode == "Blender" and not bl_crop_mesh:
+        if self.alg_mode == "Blender" and not crop_mesh_delaunay:
             return
 
         max_len = max([len(sock.sv_get(deepcopy=False)) for sock in self.inputs])
@@ -192,7 +116,7 @@ class SvCropMesh2D(bpy.types.Node, SverchCustomTreeNode):
                     out.append(crop_mesh(sv_verts, sv_faces_edges, sv_verts_crop, sv_faces_crop,
                                          self.mode, self.accuracy))
                 else:
-                    out.append(get_bl_crop_mesh_faces(sv_verts, sv_faces_edges, sv_verts_crop, sv_faces_crop,
+                    out.append(crop_mesh_delaunay(sv_verts, sv_faces_edges, sv_verts_crop, sv_faces_crop,
                                                       self.mode, 1 / 10 ** self.accuracy))
             else:
                 out.append(crop_edges(sv_verts, sv_faces_edges, sv_verts_crop, sv_faces_crop, self.mode, self.accuracy))
