+++ utils/geom_2d/merge_mesh.py
@@ -6,6 +6,13 @@
 # License-Filename: LICENSE
 
 from typing import Set, List
+from itertools import chain, cycle
+
+try:
+    from mathutils.geometry import delaunay_2d_cdt as bl_crop_mesh
+except ImportError:
+    bl_crop_mesh = None
+from mathutils import Vector
 
 from .intersections import Point as InterPoint, HalfEdge as InterHalfEdge, DCELMesh as InterDCELMesh, find_intersections
 from .make_monotone import Point as MonPoint, HalfEdge as MonHalfEdge, DCELMesh as MonDCELMesh, \
@@ -13,6 +20,7 @@ from .make_monotone import Point as MonPoint, HalfEdge as MonHalfEdge, DCELMesh 
 
 from .dcel_debugger import Debugger
 
+from sverchok.utils.geom_2d.lin_alg import is_ccw_polygon
 
 def edges_to_faces(sv_verts, sv_edges, do_intersect=True, fill_holes=True, accuracy=1e-5):
     """
@@ -268,3 +276,80 @@ def get_face_mask_by_flag(mesh, flag, del_flag='del'):
 
 def get_number_of_overlapping_mask(mesh, del_flag='del'):
     return [len(face.outer.in_faces) - 1 for face in mesh.faces if del_flag not in face.flags]
+
+def join_meshes(verts1, faces1, verts2, faces2):
+    faces_out = faces1 + [[i + len(verts1) for i in f] for f in faces2]
+    faces1_indexes = {i for i in range(len(faces1))}
+    faces2_indexes = {i + len(faces1) for i in range(len(faces2))}
+    return verts1 + verts2, faces_out, faces1_indexes, faces2_indexes
+
+
+def join_meshes2(verts1, edges1, verts2, faces2):
+    faces_out = [[i + len(verts1) for i in f] for f in faces2]
+    return verts1 + verts2, edges1, faces_out
+
+
+def del_loose(verts, poly_edge):
+    indx = set(chain.from_iterable(poly_edge))
+    verts_out = [v for i, v in enumerate(verts) if i in indx]
+    v_index = dict([(j, i) for i, j in enumerate(sorted(indx))])
+    poly_edge_out = [list(map(lambda n: v_index[n], p)) for p in poly_edge]
+    return verts_out, poly_edge_out
+
+if bl_crop_mesh is not None:
+    def crop_mesh_delaunay(verts, faces, verts_crop, faces_crop, mode, epsilon):
+        """
+        Crop given faces by another set of faces by Blender internal function
+        :param verts: list of Sv vertices
+        :param faces: list of faces wich should be cropped
+        :param verts_crop: list of Sv vertices
+        :param faces_crop: list of faces by which mesh should be cropped
+        :param mode: 'inner' or 'outer'
+        :param epsilon: float, for nearness tests
+        :return: list of Sv vertices, faces, indexes of old faces
+        """
+        faces = [f if is_ccw_polygon([verts[i] for i in f], accuracy=epsilon) else f[::-1] for f in faces]
+        faces_crop = [f if is_ccw_polygon([verts_crop[i] for i in f], accuracy=epsilon) else f[::-1] for f in faces_crop]
+        merged_verts, merged_faces, faces_indexes, faces_crop_indexes = join_meshes(verts, faces, verts_crop, faces_crop)
+        merged_verts = [Vector(co[:2]) for co in merged_verts]
+        verts_new, _, faces_new, _, _, face_indexes = bl_crop_mesh(merged_verts, [], merged_faces, 3, epsilon)
+        if mode == 'inner':
+            faces_out = []
+            faces_index_out = []
+            for f, fi in zip(faces_new, face_indexes):
+                if not fi:
+                    # it means new faces was generated
+                    continue
+                in_1 = False
+                in_2 = False
+                for i in fi:
+                    if i in faces_indexes:
+                        in_1 = True
+                    else:
+                        in_2 = True
+                    if in_1 and in_2:
+                        faces_out.append(f)
+                        faces_index_out.append(min(fi))
+                        break
+            verts_out, faces_out = del_loose(verts_new, faces_out)
+            return [v.to_3d()[:] for v in verts_out], faces_out, faces_index_out
+        else:
+            faces_out = []
+            faces_index_out = []
+            for f, fi in zip(faces_new, face_indexes):
+                if not fi:
+                    # it means new faces was generated
+                    continue
+                in_2 = False
+                for i in fi:
+                    if i in faces_crop_indexes:
+                        in_2 = True
+                        break
+                if not in_2:
+                    faces_out.append(f)
+                    faces_index_out.append(min(fi))
+            verts_out, faces_out = del_loose(verts_new, faces_out)
+            return [v.to_3d()[:] for v in verts_out], faces_out, faces_index_out
+else:
+    crop_mesh_delaunay = None
+
